requirements.txt
adlfs==2023.4.0
aiobotocore==2.5.3
aiohttp==3.8.5
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
alibi-detect==0.11.4
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argcomplete==3.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
asyncssh==2.13.2
atpublic==4.0
attrs==23.1.0
azure-core==1.28.0
azure-datalake-store==0.0.53
azure-identity==1.13.0
azure-storage-blob==12.17.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
boto3==1.28.17
botocore==1.31.17
cachetools==5.3.1
catalogue==2.0.9
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.4
configobj==5.0.8
contourpy==1.1.0
crcmod==1.7
cryptography==41.0.3
csv2md==1.1.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
dill==0.3.7
diskcache==5.6.1
distro==1.8.0
dpath==2.1.6
dulwich==0.21.5
dvc==3.13.0
dvc-azure==2.22.0
dvc-data==2.12.1
dvc-gdrive==2.20.0
dvc-gs==2.22.1
dvc-hdfs==2.19.0
dvc-http==2.30.2
dvc-objects==0.24.1
dvc-oss==2.19.0
dvc-render==0.5.3
dvc-s3==2.23.0
dvc-ssh==2.22.1
dvc-studio-client==0.12.0
dvc-task==0.3.0
dvc-webdav==2.19.1
dvc-webhdfs==2.19.0
eli5==0.13.0
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.101.0
fastjsonschema==2.18.0
filelock==3.12.2
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.42.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
funcy==2.0
gcsfs==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grandalf==0.8
graphviz==0.20.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.16.4
hydra-core==1.3.2
idna==3.4
imageio==2.31.1
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isodate==0.6.1
isoduration==20.11.0
iterative-telemetry==0.0.8
jedi==0.19.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
knack==0.11.0
kombu==5.3.1
lazy_loader==0.3
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
msal==1.23.0
msal-extensions==1.0.0
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook==7.0.2
notebook_shim==0.2.3
numba==0.57.1
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.74
orjson==3.9.3
oss2==2.18.1
ossfs==2021.8.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
plotly==5.15.0
plotly-express==0.4.1
portalocker==2.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.12
pydot==1.4.2
PyDrive2==1.16.1
pygit2==1.12.2
Pygments==2.16.1
pygtrie==2.5.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-box==7.0.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3fs==2023.6.0
s3transfer==0.6.1
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.11.1
scmrepo==1.1.0
seaborn==0.12.2
Send2Trash==1.8.2
shortuuid==1.0.11
shtab==1.6.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqltrie==0.7.0
sshfs==2023.7.0
stack-data==0.6.2
starlette==0.27.0
statsmodels==0.14.0
tabulate==0.9.0
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.7.18
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.6
webcolors==1.13
webdav4==0.9.8
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.8
wrapt==1.15.0
xgboost==1.7.6
yarl==1.9.2
zc.lockfile==3.0.post1